+++ android/versioned-abis/expoview-abi43_0_0/src/main/java/abi43_0_0/expo/modules/taskManager/TaskManagerModule.java
@@ -7,8 +7,8 @@ import abi43_0_0.expo.modules.core.ExportedModule;
 import abi43_0_0.expo.modules.core.ModuleRegistry;
 import abi43_0_0.expo.modules.core.Promise;
 import abi43_0_0.expo.modules.core.interfaces.ExpoMethod;
-import org.unimodules.interfaces.taskManager.TaskManagerInterface;
-import org.unimodules.interfaces.taskManager.TaskServiceInterface;
+import expo.modules.interfaces.taskManager.TaskManagerInterface;
+import expo.modules.interfaces.taskManager.TaskServiceInterface;
 
 import java.util.HashMap;
 import java.util.Map;