+++ packages/expo-background-fetch/android/src/main/java/expo/modules/backgroundfetch/BackgroundFetchTaskConsumer.java
@@ -10,11 +10,11 @@ import android.os.SystemClock;
 import android.util.Log;
 
 import expo.modules.core.interfaces.LifecycleEventListener;
-import org.unimodules.interfaces.taskManager.TaskConsumer;
-import org.unimodules.interfaces.taskManager.TaskConsumerInterface;
-import org.unimodules.interfaces.taskManager.TaskExecutionCallback;
-import org.unimodules.interfaces.taskManager.TaskInterface;
-import org.unimodules.interfaces.taskManager.TaskManagerUtilsInterface;
+import expo.modules.interfaces.taskManager.TaskConsumer;
+import expo.modules.interfaces.taskManager.TaskConsumerInterface;
+import expo.modules.interfaces.taskManager.TaskExecutionCallback;
+import expo.modules.interfaces.taskManager.TaskInterface;
+import expo.modules.interfaces.taskManager.TaskManagerUtilsInterface;
 
 import java.util.Map;