+++ packages/expo-updates/ios/Tests/EXUpdatesCodeSigningAlgorithmTests.swift
@@ -6,10 +6,10 @@ import XCTest
 
 class EXUpdatesCodeSigningAlgorithmTests : XCTestCase {
   func test_parseCodeSigningAlgorithm() {
-    XCTAssertEqual(try parseCodeSigningAlgorithm(nil), EXUpdatesCodeSigningAlgorithm.RSA_SHA256)
-    XCTAssertEqual(try parseCodeSigningAlgorithm("rsa-v1_5-sha256"), EXUpdatesCodeSigningAlgorithm.RSA_SHA256)
-    XCTAssertThrowsError(try parseCodeSigningAlgorithm("invalid")) { error in
-      XCTAssertEqual(error as? EXUpdatesCodeSigningAlgorithmError, EXUpdatesCodeSigningAlgorithmError.parseError)
+    XCTAssertEqual(try EXUpdatesCodeSigningAlgorithm.parseFromString(nil), EXUpdatesCodeSigningAlgorithm.RSA_SHA256)
+    XCTAssertEqual(try EXUpdatesCodeSigningAlgorithm.parseFromString("rsa-v1_5-sha256"), EXUpdatesCodeSigningAlgorithm.RSA_SHA256)
+    XCTAssertThrowsError(try EXUpdatesCodeSigningAlgorithm.parseFromString("invalid")) { error in
+      XCTAssertEqual(error as? EXUpdatesCodeSigningError, EXUpdatesCodeSigningError.AlgorithmParseError)
     }
   }
 }